back_end/requirements.txt
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
requests==2.23.0
requests-oauthlib==1.3.0
